move static-model
