+++ apps/catalog/catalog/src/app/marketplace/movie/search/search.form.ts
@@ -17,11 +17,11 @@ import {
   MovieStatusLabel,
   MovieStatusSlug,
   MOVIE_STATUS_SLUG
-} from '@blockframes/movie/movie/static-model/types';
+} from '@blockframes/utils/static-model/types';
 import { Validators, FormArray } from '@angular/forms';
 import { FormGroup, FormControl } from '@angular/forms';
 import { FormEntity, yearValidators, numberRangeValidator } from '@blockframes/utils';
-import { getLabelByCode } from '@blockframes/movie/movie/static-model/staticModels';
+import { getLabelByCode } from '@blockframes/utils/static-model/staticModels';
 import { MovieLanguageSpecification } from '@blockframes/movie/movie/+state/movie.firestore';
 import { createMovieLanguageSpecification } from '@blockframes/movie/movie+state/movie.model';
