+++ libs/movie/src/lib/movie/+state/movie.firestore.ts
@@ -13,7 +13,7 @@ import {
   FormatQualitySlug,
   FormatSlug,
   GenresSlug
-} from "@blockframes/movie/movie/static-model";
+} from "@blockframes/utils/static-model";
 import { RawRange, NumberRange } from "@blockframes/utils/common-interfaces/range";
 import { Person, SalesAgent, Company, Producer, Crew, Cast } from "@blockframes/utils/common-interfaces/identity";
 import { firestore } from "firebase/app";
