+++ libs/movie/src/lib/movie/display/main/main.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
 import { MovieMain, createMovieMain } from '../../+state';
-import { getLabelByCode } from '../../static-model/staticModels';
+import { getLabelByCode } from '@blockframes/utils/static-model/staticModels';
 
 @Component({
   selector: '[main] movie-display-main',
