+++ libs/movie/src/lib/movie/display/sales-agent-deal/sales-agent-deal.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
 import { createMovieSalesAgentDeal, MovieSalesAgentDeal } from '../../+state';
-import { getLabelByCode } from '../../static-model/staticModels';
+import { getLabelByCode } from '@blockframes/utils/static-model/staticModels';
 
 @Component({
   selector: '[deal] movie-display-salesdeal',
