--- libs/movie/src/lib/movie/form/main/origin-country/country.form.ts
@@ -1,8 +1,0 @@
-import { FormValue, isSlugValidator } from '@blockframes/utils/form';
-import { TerritoriesSlug } from '../../../static-model/types';
-
-export class FormCountry extends FormValue<TerritoriesSlug> {
-  constructor(country: TerritoriesSlug) {
-    super(country, isSlugValidator('TERRITORIES'))
-  }
-}
