+++ libs/movie/src/lib/movie/form/sales-info/original-releases/original-releases.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, Input, ChangeDetectionStrategy, OnDestroy } from '@a
 import { OriginalReleaseForm } from '../sales-info.form';
 import { MovieForm } from '../../movie.form';
 import { startWith } from 'rxjs/operators';
-import { default as staticModel, ExtractCode } from '../../../static-model/staticModels';
+import { default as staticModel, ExtractCode } from '@blockframes/utils/static-model/staticModels';
 import { MovieMainControl } from '../../main/main.form';
 import { MovieSalesInfoControl } from '../sales-info.form'
 import { Subscription } from 'rxjs';
