+++ libs/utils/src/lib/static-model/index.ts
