+++ libs/utils/src/lib/static-model/types.ts
